requirements.txt
pandas==3.0.1
plotly==6.6.0
google-genai==1.68.0
python-dotenv==1.2.2
streamlit==1.55.0
